trading/rl/rl_export/rl/requirements.txt
stable-baselines3==2.7.1
gymnasium==1.2.3
pandas-ta==0.4.71b0
python-binance==1.0.35
torch==2.10.0
numpy==2.2.6
pandas==2.3.3
matplotlib==3.10.8
